@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.6 <3.0.0'
+  sdk: '>=3.0.3 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -38,7 +38,8 @@ dependencies:
   google_sign_in: ^6.1.4
   kakao_flutter_sdk_user: ^1.4.3
   flutter_naver_login: ^1.8.0
-  http: ^0.13.5
+  sign_in_with_apple: ^4.3.0
+  http: ^1.0.0
 
 dev_dependencies:
   flutter_test:
@@ -63,7 +64,8 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
+  assets:
+    - assets/images/
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
 
